pathfinding/player/input controller working
some changes made, updated Player class, currently called NewPlayer - jeremy's is still there just need to port over whatever is needed from that.

re-implemented the pathfinding code, heap, and path request manager,

updated the game controller, to contain a World object, the world object has a reference to the tile map and list of entities, tiles in it.. extending this for items in the world too later...

currently to run this, the controllers and plot scene need to be opened together - unity's scene manager can resolve this once main menu is done...

hover tile system was updated.

Player is not in scene at start but spawned at runtime based on a supplied position - this will come from save file later.

Camera was removed as a child of player, this is so it can follow the player. needed to do this because had to change the UI to screen space - camera rather than overlay - this was to get the mouse position in world correctly and accurately.

created new class called InputController, this replaces old input coming from class called Pathfinding.cs (not to be mistaken by Pathfinder.cs)