Feature finished Inventory System
Completed the inventory system.
Ability to add new items to the game and game item dictionary.
Variable sized inventories.
No infinite inventories unless someone can tell me how to make an infinite array for some reason.
Dropping works with moving items out of the inventory but currently doesnt drop in game.
Only thing left to add is the ability to click on an item and build / use it.
Need to write a procedure on how to add new items and item catagories to the game.